message de commit de Leïna

+++ src/main/java/logic/Planete.java
@@ -1,11 +1,22 @@
 package logic;
 
+import javafx.animation.KeyFrame;
+import javafx.animation.KeyValue;
+import javafx.animation.Timeline;
+import javafx.scene.Node;
+import javafx.scene.shape.Circle;
+import javafx.scene.transform.Rotate;
+import javafx.util.Duration;
+import view.SimulationView;
+
+import static javafx.scene.paint.Color.WHITE;
+
 public class Planete extends Astre{
     private float taille;
     private float masse;
     private String nom;
-    private int positionX;
-    private int positionY;
+    private double positionX;
+    private double positionY;
 
 
     public Planete(String nom, float taille, float masse, int pX, int pY ) {
@@ -43,11 +54,11 @@ public class Planete extends Astre{
         return nom;
     }
 
-    public int getPositionX() {
+    public double getPositionX() {
         return positionX;
     }
 
-    public int getPositionY() {
+    public double getPositionY() {
         return positionY;
     }
 }
